requirements.txt
env==0.1.0
numpy==1.18.5
pandas==1.0.5
scikit_learn==1.0.2
pytorch==1.8.1
torchvision==0.9.1
tqdm==4.65.0
timm==0.5.4